fix(memoryBackend): Update testEquals for in-memory backend filters to convert String criteria to Booleans.
Allows a "true" or "false" criteria from frontend to match a boolean true or false, respectively.

It's a little "iffy" whether this should be happening at this low-level, but, justifications are:
- similar existing type coercion in this class for numbers
- similar behavior in RDBMS parameter binding
- this being closer to what any actual user would want, probably...

One could imagine that these type conversions should be happening upstream of this class.  But that does put the burden on many callers, which is a disadvantage.

One could also imagine that this behavior should be opt-in (similar to the CriteriaOption.CASE_INSENSITIVE) - so, that could be path to change this behavior in the future.  e.g., if we find a use case where someone wants a Boolean false to equal String "false".  So we'd probably wrap the existing type-converting behavior in here in something like: if(!criterion.hasOption(CriteriaOption.STRICT_TYPES)) or .NO_TYPE_CONVERSIONS